extracted inventory activate handler
